requirements.txt
black==22.3.0
Bottleneck==1.3.4
certifi==2021.10.8
click==8.1.3
cycler==0.11.0
fonttools==4.25.0
importlib-metadata==4.11.3
kiwisolver==1.3.2
matplotlib==3.5.1
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
munkres==1.1.4
mypy-extensions==0.4.3
numexpr==2.8.1
numpy==1.21.5
packaging==21.3
pandas==1.3.4
pathspec==0.9.0
Pillow==9.0.1
pip==22.0.4
platformdirs==2.5.2
pyparsing==3.0.4
python-dateutil==2.8.2
pytz==2021.3
scipy==1.7.3
seaborn==0.11.2
setuptools==61.2.0
six==1.16.0
tomli==2.0.1
tornado==6.1
typed-ast==1.5.3
typing_extensions==4.2.0
wheel==0.37.1
zipp==3.8.0
